refactor : CartStorage의 자료구조 변경 및 CartRepository 인터페이스 수정
- CartStorage에서 Product 대신 Cart를 저장하도록 변경
- CartStorage의 cart 자료구조를 MutableSet에서 LinkedHashMap으로 변경하여 ID를 통한 접근이 가능하도록 수정
- CartRepository 인터페이스의 insert 메서드 파라미터를 Product에서 Long 타입의 productId로 변경
- CartRepository 인터페이스에 get 메서드를 추가하여 ID로 Cart를 조회할 수 있도록 기능 추가
- CartRepositoryImpl에서 위 변경사항들을 반영
