+++ app/src/main/java/woowacourse/shopping/data/repository/CartRepositoryImpl.kt
@@ -1,14 +1,16 @@
 package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.data.storage.CartStorage
-import woowacourse.shopping.domain.CartResult
-import woowacourse.shopping.domain.product.Product
+import woowacourse.shopping.domain.cart.Cart
+import woowacourse.shopping.domain.cart.CartResult
 import woowacourse.shopping.domain.repository.CartRepository
 
 class CartRepositoryImpl(private val storage: CartStorage) : CartRepository {
-    override fun insert(item: Product) = storage.insertProduct(item)
+    override fun insert(productId: Long) = storage.insert(Cart(productId = productId))
 
-    override fun delete(id: Long) = storage.deleteProduct(id)
+    override fun get(id: Long): Cart = storage[id]
+
+    override fun delete(id: Long) = storage.delete(id)
 
     override fun loadSinglePage(
         page: Int,
